Move __getstate__/__setstate__ to the right Popen classes
