+++ Lib/multiprocessing/popen_spawn_win32.py
@@ -129,3 +129,13 @@ class Popen(object):
 
     def close(self):
         self.finalizer()
+
+    def __getstate__(self):
+        state = vars(self).copy()
+        state['finalizer'] = state['finalizer']._key
+        return state
+
+    def __setstate__(self, state):
+        vars(self).update(state)
+        self.finalizer = util._finalizer_registry[self.finalizer]
+
