+++ Lib/multiprocessing/process.py
@@ -285,16 +285,6 @@ class BaseProcess(object):
             info.append('daemon')
         return '<%s>' % ' '.join(info)
 
-    def __getstate__(self):
-        state = vars(self).copy()
-        state['finalizer'] = state['finalizer']._key
-        return state
-
-    def __setstate__(self, state):
-        from . import util
-        vars(self).update(state)
-        self.finalizer = util._finalizer_registry[self.finalizer]
-
     ##
 
     def _bootstrap(self, parent_sentinel=None):
